Fixed issue with WebSocket message broadcasting
This commit fixes the issue where WebSocket messages were not being broadcasted to all connected clients. The problem was caused by the `clientId` is the client's identifier, so if it is removed, the system won't be able to identify which client joined the room, sent a particular message, or leaved the room.